refactor(buttons): move button dir into animate-ui dir
