+++ src/components/LiquidButton.tsx
@@ -1,21 +1,28 @@
 import type { IconType } from "react-icons";
-import { LiquidButton as AnimatedButton } from "./buttons/liquid";
+import { LiquidButton as AnimatedButton } from "./animate-ui/buttons/liquid";
 
 interface LiquidButtonProps {
-Icon: IconType;
-text: string;
-onClick?: React.MouseEventHandler<HTMLButtonElement>;
-link?: string;
+	Icon: IconType;
+	text: string;
+	onClick?: React.MouseEventHandler<HTMLButtonElement>;
+	link?: string;
 }
-const LiquidButton = ({Icon, text, onClick, link }: LiquidButtonProps) => {
-return (
-    <AnimatedButton onClick={onClick} className="group grow">
-    <a href={ link } target="_blank" className="flex gap-2 text-foreground items-center">
-     <Icon className="text-primary group-hover:text-foreground delay-300" style={{width: "20px", height: "20px"}}/> 
-        {text}
-      </a>
-     </AnimatedButton>  
-  )
-}
+const LiquidButton = ({ Icon, text, onClick, link }: LiquidButtonProps) => {
+	return (
+		<AnimatedButton onClick={onClick} className="group grow">
+			<a
+				href={link}
+				target="_blank"
+				className="flex gap-2 text-foreground items-center"
+			>
+				<Icon
+					className="text-primary group-hover:text-foreground delay-300"
+					style={{ width: "20px", height: "20px" }}
+				/>
+				{text}
+			</a>
+		</AnimatedButton>
+	);
+};
 
-export default LiquidButton
+export default LiquidButton;
