+++ src/components/animate-ui/buttons/liquid.tsx
@@ -3,7 +3,7 @@
 import { motion, type HTMLMotionProps } from "motion/react";
 import { cva, type VariantProps } from "class-variance-authority";
 
-import { cn } from "../../lib/utils";
+import { cn } from "../../../lib/utils";
 
 const buttonVariants = cva(
 	"relative inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-lg text-sm font-medium cursor-pointer overflow-hidden disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive [background:_linear-gradient(var(--liquid-button-color)_0_0)_no-repeat_calc(200%-var(--liquid-button-fill,0%))_100%/200%_var(--liquid-button-fill,0.2em)] hover:[--liquid-button-fill:100%] hover:[--liquid-button-delay:0.3s] [transition:_background_0.3s_var(--liquid-button-delay,0s),_color_0.3s_var(--liquid-button-delay,0s),_background-position_0.3s_calc(0.3s_-_var(--liquid-button-delay,0s))] focus:outline-none",
@@ -14,7 +14,7 @@ const buttonVariants = cva(
 					"text-foreground hover:text-background-dark !bg-muted [--liquid-button-color:var(--primary)]",
 				outline:
 					"border !bg-background dark:!bg-input/30 dark:border-input [--liquid-button-color:var(--primary)]",
-				secondary: 
+				secondary:
 					"text-secondary hover:text-secondary-foreground !bg-muted [--liquid-button-color:var(--secondary)]",
 			},
 			size: {
